ElementCount field
The new field, ElementCount, is a subtype of CalculatedInt.
It holds the count of direct sub-field of a container, or, in the case
of a field - 1. As with other calculated fields, the depends_on field
may be referred to by object or by name. Also, a user may provide a
custom function to calculate the value of the field.

A new API was introduced to BaseField, get_rendered_fields, which
returned an ordered list of the currently rendered fields.