[DataModel] new field - Float
Float is a field that represents an IEEE 754 floating point number.
At this point, the mutation library for it only contains special cases
(e.g. negative and positives of inf, Nan and 0).
A set of Float encoders was added as well, based on struct (for binary32
and binary64) and format strings (for decimal*).