[DataModel] bugfix: Container is_default did not check self mutations
the is_default function of the container only checked if all sub-fields
are with default values. However, some containers have mutations even
when the subfields are with default values, and in those cases the wrong
output was rendered.

For example:

    Container(
        fields=Repeat(
            fields=Static('a'), min_times=2, max_times=10
        )
    )

Always rendered to the same value, even when mutated (instead of
repeating).

Thanks @dark-lbp for raising this issue.