[TestList] new module to handle the list of tests to be run
This module is part of the preparation to close issue #26:
"Add kitty-option: validate - to rerun all the failed/error tests".

This module contains two classes ATM:
1. StartEndList - this is basically the test list that we had so far, it
points to tests from the start index to the end index
2. RangesList - this list is more flexible, and can point to non
consecutive test indices (still sorted though). Eventually this class
will succeed StartEndList and make it obsolete, but for now it is only
an untested skeleton. This allows a user to specify easily a list of non
consecutive tests or single tests as well as will allow kitty to support
re-run of failed tests automatically.