[Fuzzer] add "retest" option
A new command line option "--retest sessionfile" instructs the fuzzer to
run only the list of tests that have report in the session file.
It will not change the sessionfile, only use it to get the information
about the reports, if storing the information about the retest session
is required - it could still be achieved with "--session-file
sessionfile".

This commit closes issue #26 - "Add kitty-option: validate - to rerun
all the failed/error tests".