[DataModel] bugfix: set default sequence for StagedSequenceModel
This commit fixes issue #62 - "StagedSequenceModel bug"
It adds a default template sequence to the data model so that this
high-level model will provide a template sequence even if it is not
mutated.

The test that was provided with the issue exposed a couple of additional
issues that were fixed as well, and the unit tests were modified to deal
with the changes in this commit.