[Fuzzer] Optimize test list management in the fuzzer
This commit closes issue #64 - "[Fuzzer] Optimize test list manager so
it won't hold long lists of tests"

Instead of having a list of all the tests that should run, which might
consume a lot of memory, the RangesList now uses StartEndList internally
and so reduces its memory consumption.

This commit contains partial refactoring of the entire test_list module.