[DataModel] Fixed default value of Pad container
Default rendered value of the Pad was not calculated properly, and it
was not padded. This caused an issue when the Pad was not mutated and
render was called on its enclosing container, as it did not render the
Pad, only took its default value.

This commit fixes issue #66 - "Default value of Pad field is not padded"