[DataModel] List - raise exception if no fields are passed
List requires that its fields will be passed in the fields argument of
the constructor and not by calling push, otherwise the List will
function as OneOf, which helps nobody. To avoid cases where this happens
by mistake, the List constructor will now raise an exception if an empty
list or None is passed to List.
This commit contains test cases as well.