Serverfuzzer: call edge.callback() after receiving response
Before, the edge.callback function would be called before the response was received, when it was the first edge in the sequence, which meant the callback function would be called with None as response instead of the real response. This commit moves the callback call down so the first response in the sequence will be the correct one and not None.